Renamed module awsJobStore.__init__ to aws.jobStore
